# bt: Implementing binary tree backed by a vector.