Related to #947.

Using a vector of nodes to represent an append-only binary tree.

E.g. This is the representation of a tree with three nodes. 
```mermaid
flowchart TB
    A --> B
    A --> C
```
In-memory storage:

|NodeRef| Node | Left | Right |
|--|--|--|--|
|0| A | 1 | 2|
|1| B | None  | None |
|2| C | None  | None |

Encoded representation: Nodes are serialized in a pre-order ordering.

The table of nodes starts with `NODES_CAPACITY=256` nodes pre-allocated in order to reduce frequent allocations.

## Benchmark Comparison

Code for the benchmark comparison: https://github.com/armfazh/libprio-rs/pull/3

**Task:** Insert `NUM_PATHS=1000` random paths of length N.

Old: Uses canonical pointers data structure (main).
New: Uses a table (`Vec`) to store nodes (this PR).

Root: Traverses the tree from the root node.
Node: Traverses the tree from the last node inserted. 

See **new** timings [below](https://github.com/divviup/libprio-rs/pull/1080#issuecomment-2225410553) in the thread of comments.