@simon-friedberger 2024-07-02
I think this code deserves a comment on why a generic binary tree from a crate wouldn't do or why this shouldn't be a separate crate. It looks like a fairly generic interface to have a binary tree without node removal.