@armfazh 2024-07-10
I updated the description of the PR (top). Now it includes timings that compare the current implementation with the implementation in this PR. 

The savings are more significant when using large paths (i.e. large tree's height). Also, insertion in-place offers significant savings too. 

The code for benchmark is [here](https://github.com/armfazh/libprio-rs/pull/3).