@cjpatton 2024-07-10
> The savings are more significant when using large paths (i.e. large tree's height). Also, insertion in-place offers significant savings too.

That doesn't seem to be corroborated by the timings you posted:


> `N=8`
> 
> | Insertion at | Old | New |
> |--|--|--|
> | Root | 199.47 µs | 185.05 µs | 
> | Node | 235.31 µs | 180.04 µs | 
> 
> `N=64`
> 
> | Insertion at | Old | New |
> |--|--|--|
> | Root | 5.6078 ms | 5.3975 ms | 
> | Node | 2.8733 ms | 1.0842 ms | 
> 
> `N=256`
> 
> | Insertion at | Old | New |
> |--|--|--|
> | Root | 66.765 ms | 93.256 ms | 
> | Node | 11.680 ms |  4.219 ms | 

From this table it looks like inserting from the root is significantly slower for the new code.