@branlwyd 2024-07-11
It appears that insert-from-root only becomes slower once `N` becomes large enough. If `N` continues to increase, does the performance difference become larger?