@armfazh 2024-07-12
I revisited the benchmarks and updated the bench script to compute the task correctly.

As @branlwyd pointed out, the time for inserting values consistently increases as the length of paths grows.

With these new measurements, it's clear that the new version (using `std::Vec`) is no better than the current code. 
One reason that explains this is because the expression `self.nodes[node]` is a runtime checked (unlike in C). I tried to use the `unsafe get_unchecked` method but only gives 15% savings, it's not closer to the current code.

**N** is length of path.

### Inserting at Root

| N | Old (Tree) | New (Vector) |
|--|--|--|
|  16  | 293.27 µs | 353.73 µs |
|  64  |   3.56 ms |   5.21 ms |
| 128  |  14.46 ms |  22.05 ms |
| 256  |  58.81 ms |  91.52 ms |
| 512  | 239.45 ms | 374.53 ms |

### Inserting at Node

| N | Old (Tree) | New (Vector) |
|--|--|--|
|  16  | 172.80 µs | 212.11 µs |
|  64  | 686.39 µs | 853.18 µs |
| 128  | 1.3668 ms | 1.6735 ms |
| 256  | 2.7321 ms | 3.3498 ms |
| 512  | 5.4535 ms | 6.6705 ms |